+++ packages/flutter/test/screenshot/masking_config_test.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
@@ -125,7 +126,7 @@ void main() async {
           SentryMaskingDecision.unmask);
     });
 
-    testWidgets('retuns false if no rule matches', (tester) async {
+    testWidgets('returns false if no rule matches', (tester) async {
       final sut = SentryMaskingConfig([
         SentryMaskingCustomRule<Image>(
           callback: (e, w) => SentryMaskingDecision.continueProcessing,
@@ -169,6 +170,7 @@ void main() async {
         'SentryMaskingConstantRule<Text>(mask)',
         'SentryMaskingConstantRule<EditableText>(mask)',
         'SentryMaskingConstantRule<RichText>(mask)',
+        ..._maybeWithSensitiveContent(),
         'SentryMaskingCustomRule<Widget>(Debug-mode-only warning for potentially sensitive widgets.)'
       ]);
     });
@@ -181,6 +183,7 @@ void main() async {
       expect(rulesAsStrings(sut), [
         ...alwaysEnabledRules,
         'SentryMaskingConstantRule<Image>(mask)',
+        ..._maybeWithSensitiveContent(),
         'SentryMaskingCustomRule<Widget>(Debug-mode-only warning for potentially sensitive widgets.)'
       ]);
     });
@@ -193,6 +196,7 @@ void main() async {
       expect(rulesAsStrings(sut), [
         ...alwaysEnabledRules,
         'SentryMaskingCustomRule<Image>(Mask all images except asset images.)',
+        ..._maybeWithSensitiveContent(),
         'SentryMaskingCustomRule<Widget>(Debug-mode-only warning for potentially sensitive widgets.)'
       ]);
     });
@@ -207,6 +211,7 @@ void main() async {
         'SentryMaskingConstantRule<Text>(mask)',
         'SentryMaskingConstantRule<EditableText>(mask)',
         'SentryMaskingConstantRule<RichText>(mask)',
+        ..._maybeWithSensitiveContent(),
         'SentryMaskingCustomRule<Widget>(Debug-mode-only warning for potentially sensitive widgets.)'
       ]);
     });
@@ -218,9 +223,36 @@ void main() async {
         ..maskAssetImages = false;
       expect(rulesAsStrings(sut), [
         ...alwaysEnabledRules,
+        ..._maybeWithSensitiveContent(),
         'SentryMaskingCustomRule<Widget>(Debug-mode-only warning for potentially sensitive widgets.)'
       ]);
     });
+
+    test(
+        'SensitiveContent rule is automatically added when current Flutter version is equal or newer than 3.33',
+        () {
+      final sut = SentryPrivacyOptions();
+      final version = FlutterVersion.version!;
+      final dot = version.indexOf('.');
+      final major = int.tryParse(version.substring(0, dot));
+      final nextDot = version.indexOf('.', dot + 1);
+      final minor = int.tryParse(
+          version.substring(dot + 1, nextDot == -1 ? version.length : nextDot));
+
+      if (major! > 3 || (major == 3 && minor! >= 33)) {
+        expect(
+            rulesAsStrings(sut).contains(
+                'SentryMaskingCustomRule<SensitiveContent>(Mask SensitiveContent widget.)'),
+            isTrue,
+            reason: 'Test failed with version: ${FlutterVersion.version}');
+      } else {
+        expect(
+            rulesAsStrings(sut).contains(
+                'SentryMaskingCustomRule<SensitiveContent>(Mask SensitiveContent widget.)'),
+            isFalse,
+            reason: 'Test failed with version: ${FlutterVersion.version}');
+      }
+    }, skip: FlutterVersion.version == null);
 
     group('user rules', () {
       final defaultRules = [
@@ -229,20 +261,28 @@ void main() async {
         'SentryMaskingConstantRule<Text>(mask)',
         'SentryMaskingConstantRule<EditableText>(mask)',
         'SentryMaskingConstantRule<RichText>(mask)',
+        ..._maybeWithSensitiveContent(),
         'SentryMaskingCustomRule<Widget>(Debug-mode-only warning for potentially sensitive widgets.)'
       ];
+
       test('mask() takes precedence', () {
         final sut = SentryPrivacyOptions();
         sut.mask<Image>();
-        expect(rulesAsStrings(sut),
-            ['SentryMaskingConstantRule<Image>(mask)', ...defaultRules]);
+        expect(rulesAsStrings(sut), [
+          'SentryMaskingConstantRule<Image>(mask)',
+          ...defaultRules,
+        ]);
       });
+
       test('unmask() takes precedence', () {
         final sut = SentryPrivacyOptions();
         sut.unmask<Image>();
-        expect(rulesAsStrings(sut),
-            ['SentryMaskingConstantRule<Image>(unmask)', ...defaultRules]);
+        expect(rulesAsStrings(sut), [
+          'SentryMaskingConstantRule<Image>(unmask)',
+          ...defaultRules,
+        ]);
       });
+
       test('are ordered in the call order', () {
         var sut = SentryPrivacyOptions();
         sut.mask<Image>();
@@ -272,13 +312,14 @@ void main() async {
           ...defaultRules
         ]);
       });
+
       test('maskCallback() takes precedence', () {
         final sut = SentryPrivacyOptions();
         sut.maskCallback(
             (Element element, Image widget) => SentryMaskingDecision.mask);
         expect(rulesAsStrings(sut), [
           'SentryMaskingCustomRule<Image>(Custom callback-based rule (description unspecified))',
-          ...defaultRules
+          ...defaultRules,
         ]);
       });
       test('User cannot add $SentryMask and $SentryUnmask rules', () {
@@ -318,6 +359,25 @@ void main() async {
     // ignored and thus no warning is logged.
     expect(logger.items.where((i) => i.level == SentryLevel.warning), isEmpty);
   });
+}
+
+List<String> _maybeWithSensitiveContent() {
+  final version = FlutterVersion.version;
+  if (version == null) {
+    return [];
+  }
+  final dot = version.indexOf('.');
+  final major = int.tryParse(version.substring(0, dot));
+  final nextDot = version.indexOf('.', dot + 1);
+  final minor = int.tryParse(
+      version.substring(dot + 1, nextDot == -1 ? version.length : nextDot));
+  if (major! > 3 || (major == 3 && minor! >= 33)) {
+    return [
+      'SentryMaskingCustomRule<SensitiveContent>(Mask SensitiveContent widget.)'
+    ];
+  } else {
+    return [];
+  }
 }
 
 extension on Element {